backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.3.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dimod==0.12.14
diskcache==5.6.3
dwave-cloud-client==0.11.2
dwave-inspector==0.4.4
dwave-networkx==0.8.14
dwave-preprocessing==0.6.5
dwave-samplers==1.2.0
dwave-system==1.23.0
exceptiongroup==1.2.0
fastapi==0.104.1
fasteners==0.19
Flask==3.0.2
h11==0.14.0
homebase==1.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
minorminer==0.2.13
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
plucky==0.4.3
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
rectangle-packer==2.0.2
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
zipp==3.17.0
